--- src/lib/react-dom/index.js
@@ -1,5 +1,0 @@
-import { render } from "./render";
-
-export const ReactDOM = {
-  render,
-};
